requirements.txt
python-telegram-bot[all]>=21.0
python-dotenv>=1.0.0
gemini-transcriber @ file:///C:/Users/soumy/projects/gemini-transcriber
